requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.27
paho-mqtt>=2.1
jinja2>=3.1
python-multipart>=0.0.20
